Added image registration tests
Added registration tests to simply test the args being passed
to the `openstack_identity_register` LWRP.  The identity LWRP
should be tested at a deeper level than the consumers.

Testing args passed to a DSL (&block) is a bit tricky.  The tests
are not ideal, but give us the necessary coverage to catch common
errors.

Change-Id: I41c5f609582f8ceb46ea73788a0a37d557532444