Added the new rabbit `["openstack"]["mq"]` attributes
After reading over the configuration, rabbit host was being
injected into the templates, however, the `notifier_strategy`
is set to `noop`.  Since mq is not configured in glance,
set the `rabbit_host` flag to the template's default value.
When someone is ready to come along and configure glance w/
rabbit, the attributes are setup properly.

Change-Id: I52f3c2b54e3313c48df0f2f0edbdeba72a526171