Making glance service names consistent
Right now for most cookbooks we declare resources with the name of the
service itself. So, for example, in the 'identity' cookbook, the
'keystone' service is called 'keystone'. The other cookboks follow this
same standard.

Why is this important? Things like omnibus-openstack depend on these
cookbooks having a common standard. That, and well, standards make us
all happy.

Change-Id: I3b1cb845cfb21b5e1d46feca322f3c2788431a27