Change execution place of ressource execute['glance-manage version_control 0'] in recipes/registry.rb
The ressource execute['glance-manage version_control 0'] crash when you change the glance db password
because the ressource template['/etc/glance/glance-registry.conf'] is not yet updated with the new password at the time.
This fix change the order of execution of the ressource execute['glance-manage version_control 0'] to fix that issue.

Change-Id: I3376c8ad4da90bef22c595945a984233f22222fb
Closes-Bug: #1282941