Changes related to increase IP binding flexibility
Changes have been made to the common cookbook so that if a
bind_interface is specified for an endpoint, then the IP is looked up
for that endpoint using address_for, and if no bind_interface is
specified then the value of the host attribute of the endpoint is used
instead.

This change set removes code that implements this exact behavior from
the image cookbook since it now is provided by the common cookbook.

blueprint increase-ip-binding-flexibility

Change-Id: I7fda0b50704e530f4de0cfc0e3a6c90f0db53e2e