Install missing dependency and override config
Installing openstack-glance installs python-openstack which pulls in
python-kombu, but python-kombu doesn't pull in the necessary
python-anyjson package.  This prevents glance-api from starting.

Additionally, openstack-glance on CentOS loads
/usr/share/glance/glance-api-dist.conf which is using the deprecated
notifier_strategy config option which defaults to qpid.  When we run
with a notification_driver != 'qpid' there seems to be some sort of
conflict and glance-api service doesn't operate correctly.

I'll be reporting both of these issues upstream, but for the time being
we need to pull in this package so we can progress w/ Icehouse on CentOS
testing.

Change-Id: Iaac29d6f55171536a8ee5cb786f49482ef51b69f
Closes-Bug: #1307358