Ensure glance-manage commands are run as correct user/group
Whilst testing stackforge cookbooks w/ ctracey's
cookbook-openstack-omnibus cookbook, I noticed that
/var/log/glance/glance-api.log was getting created as root and this
prevented glance-api from starting as it could not write to the log.

This is a non-issue when using packages as on Ubuntu (at least)
glance-manage db_sync is executed as part of glance-registry's
post-install script as user glance which ensures log file is created
with appropriate ownership.

Change-Id: I56d3e8caa68a32e34baa628a87c5e41b49602c59
Closes-Bug: #1317493