Change image_upload spec to not test the LWRP
Before this change, the image_upload recipe was kind of used to test the
LWRP that is defined by this cookbook.  Now, it only checks the
resources that appear in that recipe.

To make that work, a libraries/matchers.rb was added that helps ChefSpec
cope with the dash in the cookbook name.

Change-Id: I632556c9924d1be946b033a8759c0f3de15d91e0