Fix image upload for tar kernel/initrd images
Current logic is parsing output from glance to get the 'id' of the
kernel and initrd images.  Glance command output format was changed long ago
to a table style output and broke this.
There is already a todo in the code to refactor this but I did not attempt
that here.  Also add a todo for more provider method testers.

Change-Id: I68affe1d7f3df2ee69717641f5be6ba16615654e
Closes-Bug: #1373614