Add support for vhd vmdk vdi iso raw disk formats
Image upload provider already has a image type, but was not
used to handle the other common image types that have a simple
bare container format.   Expanded the list of types and allowed
the existing upload_image recipe to also take advantage of this
new provider disk type support in an optional way.  Backward
compatible.

Change-Id: I1f5983150aa648837c9bcd256f0ceb4bf73d4b52
Closes-Bug: #1383079