Only admin can create public glance images
With change to use the correct "service" role for service users,
they can no longer create public images.
For this recipe, need to use admin for public images.

Added a public flag to the lwrp such that other non-admin
accounts can create images.   Made a note in the client cookbook
patch that this support needs to be merged in there.

Change-Id: I99e2febfdbf6f4bab260d897216f4ae768cf3315
Related-Bug: #1436050
Closes-Bug: #1441292