Use correct endpoint scheme for api to registry connection
The ['openstack']['image']['ssl']['enabled'] was being used for
both the API/Regsitry services AND the APIs internal connection
to the registry.

Need to allow for separate control for enabling SSL for API and Registry.
Added attributes to cover this.

And the APIs internal connection to the registry needs to be
configured correctly based upon the Registry's internal endpoint scheme
(http or https).

Change-Id: Ieed28ad891192d81356c0d40ba48d70517087950
Closes-Bug: #1462388